move dayjs to utils folder
